backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# OCR and Image Processing
pytesseract==0.3.10
opencv-python==4.8.1.78
pdf2image==1.16.3
Pillow==10.1.0

# ML/AI
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
accelerate==0.25.0
bitsandbytes==0.41.1
langchain==0.0.340

# Vector Database (FAISS вместо ChromaDB)
faiss-cpu==1.7.4

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.1